requirements.txt
requests
streamlit
plotly
python-dateutil